// almora-shop-ui/src/gql/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Asset = Node & {
  __typename?: "Asset";
  createdAt: Scalars["DateTime"]["output"];
  customFields?: Maybe<Scalars["JSON"]["output"]>;
  fileSize: Scalars["Int"]["output"];
  focalPoint?: Maybe<Coordinate>;
  height: Scalars["Int"]["output"];
  id: Scalars["ID"]["output"];
  mimeType: Scalars["String"]["output"];
  name: Scalars["String"]["output"];
  preview: Scalars["String"]["output"];
  source: Scalars["String"]["output"];
  tags: Array<Tag>;
  type: AssetType;
  updatedAt: Scalars["DateTime"]["output"];
  width: Scalars["Int"]["output"];
};
// ...
export type Channel = Node & {
  __typename?: "Channel";
  availableCurrencyCodes: Array<CurrencyCode>;
  availableLanguageCodes?: Maybe<Array<LanguageCode>>;
  code: Scalars["String"]["output"];
  createdAt: Scalars["DateTime"]["output"];
  /** @deprecated Use defaultCurrencyCode instead */
  currencyCode: CurrencyCode;
  customFields?: Maybe<Scalars["JSON"]["output"]>;
  defaultCurrencyCode: CurrencyCode;
  defaultLanguageCode: LanguageCode;
  defaultShippingZone?: Maybe<Zone>;
  defaultTaxZone?: Maybe<Zone>;
  id: Scalars["ID"]["output"];
  /** Not yet used - will be implemented in a future release. */
  outOfStockThreshold?: Maybe<Scalars["Int"]["output"]>;
  pricesIncludeTax: Scalars["Boolean"]["output"];
  seller?: Maybe<Seller>;
  token: Scalars["String"]["output"];
  /** Not yet used - will be implemented in a future release. */
  trackInventory?: Maybe<Scalars["Boolean"]["output"]>;
  updatedAt: Scalars["DateTime"]["output"];
};

export type Collection = Node & {
  __typename?: "Collection";
  assets: Array<Asset>;
  breadcrumbs: Array<CollectionBreadcrumb>;
  children?: Maybe<Array<Collection>>;
  createdAt: Scalars["DateTime"]["output"];
  customFields?: Maybe<Scalars["JSON"]["output"]>;
  description: Scalars["String"]["output"];
  featuredAsset?: Maybe<Asset>;
  filters: Array<ConfigurableOperation>;
  id: Scalars["ID"]["output"];
  languageCode?: Maybe<LanguageCode>;
  name: Scalars["String"]["output"];
  parent?: Maybe<Collection>;
  parentId: Scalars["ID"]["output"];
  position: Scalars["Int"]["output"];
  productVariants: ProductVariantList;
  slug: Scalars["String"]["output"];
  translations: Array<CollectionTranslation>;
  updatedAt: Scalars["DateTime"]["output"];
};
// ...
export type CollectionList = PaginatedList & {
  __typename?: "CollectionList";
  items: Array<Collection>;
  totalItems: Scalars["Int"]["output"];
};
// ...
export type Country = Node &
  Region & {
    __typename?: "Country";
    code: Scalars["String"]["output"];
    createdAt: Scalars["DateTime"]["output"];
    customFields?: Maybe<Scalars["JSON"]["output"]>;
    enabled: Scalars["Boolean"]["output"];
    id: Scalars["ID"]["output"];
    languageCode: LanguageCode;
    name: Scalars["String"]["output"];
    parent?: Maybe<Region>;
    parentId?: Maybe<Scalars["ID"]["output"]>;
    translations: Array<RegionTranslation>;
    type: Scalars["String"]["output"];
    updatedAt: Scalars["DateTime"]["output"];
  };
// ...
export type Customer = Node & {
  __typename?: "Customer";
  addresses?: Maybe<Array<Address>>;
  createdAt: Scalars["DateTime"]["output"];
  customFields?: Maybe<Scalars["JSON"]["output"]>;
  emailAddress: Scalars["String"]["output"];
  firstName: Scalars["String"]["output"];
  id: Scalars["ID"]["output"];
  lastName: Scalars["String"]["output"];
  orders: OrderList;
  phoneNumber?: Maybe<Scalars["String"]["output"]>;
  title?: Maybe<Scalars["String"]["output"]>;
  updatedAt: Scalars["DateTime"]["output"];
  user?: Maybe<User>;
};
// ...
export type Order = Node & {
  __typename?: "Order";
  /** An order is active as long as the payment process has not been completed */
  active: Scalars["Boolean"]["output"];
  billingAddress?: Maybe<OrderAddress>;
  /** A unique code for the Order */
  code: Scalars["String"]["output"];
  /** An array of all coupon codes applied to the Order */
  couponCodes: Array<Scalars["String"]["output"]>;
  createdAt: Scalars["DateTime"]["output"];
  currencyCode: CurrencyCode;
  customFields?: Maybe<Scalars["JSON"]["output"]>;
  customer?: Maybe<Customer>;
  discounts: Array<Discount>;
  fulfillments?: Maybe<Array<Fulfillment>>;
  history: HistoryEntryList;
  id: Scalars["ID"]["output"];
  lines: Array<OrderLine>;
  /**
   * The date & time that the Order was placed, i.e. the Customer
   * completed the checkout and the Order is no longer "active"
   */
  orderPlacedAt?: Maybe<Scalars["DateTime"]["output"]>;
  payments?: Maybe<Array<Payment>>;
  /** Promotions applied to the order. Only gets populated after the payment process has completed. */
  promotions: Array<Promotion>;
  shipping: Scalars["Money"]["output"];
  shippingAddress?: Maybe<OrderAddress>;
  shippingLines: Array<ShippingLine>;
  shippingWithTax: Scalars["Money"]["output"];
  state: Scalars["String"]["output"];
  /**
   * The subTotal is the total of all OrderLines in the Order. This figure also includes any Order-level
   * discounts which have been prorated (proportionally distributed) amongst the items of each OrderLine.
   * To get a total of all OrderLines which does not account for prorated discounts, use the
   * sum of `OrderLine.discountedLinePrice` values.
   */
  subTotal: Scalars["Money"]["output"];
  /** Same as subTotal, but inclusive of tax */
  subTotalWithTax: Scalars["Money"]["output"];
  /**
   * Surcharges are arbitrary modifications to the Order total which are neither
   * ProductVariants nor discounts resulting from applied Promotions. For example,
   * one-off discounts based on customer interaction, or surcharges based on payment
   * methods.
   */
  surcharges: Array<Surcharge>;
  /** A summary of the taxes being applied to this Order */
  taxSummary: Array<OrderTaxSummary>;
  /** Equal to subTotal plus shipping */
  total: Scalars["Money"]["output"];
  totalQuantity: Scalars["Int"]["output"];
  /** The final payable amount. Equal to subTotalWithTax plus shippingWithTax */
  totalWithTax: Scalars["Money"]["output"];
  type: OrderType;
  updatedAt: Scalars["DateTime"]["output"];
};
// ...
export type PaymentMethodQuote = {
  __typename?: "PaymentMethodQuote";
  code: Scalars["String"]["output"];
  customFields?: Maybe<Scalars["JSON"]["output"]>;
  description: Scalars["String"]["output"];
  eligibilityMessage?: Maybe<Scalars["String"]["output"]>;
  id: Scalars["ID"]["output"];
  isEligible: Scalars["Boolean"]["output"];
  name: Scalars["String"]["output"];
};
// ...
export type Product = Node & {
  __typename?: "Product";
  assets: Array<Asset>;
  collections: Array<Collection>;
  createdAt: Scalars["DateTime"]["output"];
  customFields?: Maybe<Scalars["JSON"]["output"]>;
  description: Scalars["String"]["output"];
  enabled: Scalars["Boolean"]["output"];
  facetValues: Array<FacetValue>;
  featuredAsset?: Maybe<Asset>;
  id: Scalars["ID"]["output"];
  languageCode: LanguageCode;
  name: Scalars["String"]["output"];
  optionGroups: Array<ProductOptionGroup>;
  slug: Scalars["String"]["output"];
  translations: Array<ProductTranslation>;
  updatedAt: Scalars["DateTime"]["output"];
  /** Returns a paginated, sortable, filterable list of ProductVariants */
  variantList: ProductVariantList;
  /** Returns all ProductVariants */
  variants: Array<ProductVariant>;
};
// ...
export type ProductVariant = Node & {
  __typename?: "ProductVariant";
  assets: Array<Asset>;
  createdAt: Scalars["DateTime"]["output"];
  currencyCode: CurrencyCode;
  customFields?: Maybe<Scalars["JSON"]["output"]>;
  facetValues: Array<FacetValue>;
  featuredAsset?: Maybe<Asset>;
  id: Scalars["ID"]["output"];
  languageCode: LanguageCode;
  name: Scalars["String"]["output"];
  options: Array<ProductOption>;
  price: Scalars["Money"]["output"];
  priceWithTax: Scalars["Money"]["output"];
  product: Product;
  productId: Scalars["ID"]["output"];
  sku: Scalars["String"]["output"];
  stockLevel: Scalars["String"]["output"];
  taxCategory: TaxCategory;
  taxRateApplied: TaxRate;
  translations: Array<ProductVariantTranslation>;
  updatedAt: Scalars["DateTime"]["output"];
};
// ...
export type PublicPaymentMethod = {
  __typename?: "PublicPaymentMethod";
  code: Scalars["String"]["output"];
  customFields?: Maybe<Scalars["JSON"]["output"]>;
  description?: Maybe<Scalars["String"]["output"]>;
  id: Scalars["ID"]["output"];
  name: Scalars["String"]["output"];
  translations: Array<PaymentMethodTranslation>;
};

export type PublicShippingMethod = {
  __typename?: "PublicShippingMethod";
  code: Scalars["String"]["output"];
  customFields?: Maybe<Scalars["JSON"]["output"]>;
  description?: Maybe<Scalars["String"]["output"]>;
  id: Scalars["ID"]["output"];
  name: Scalars["String"]["output"];
  translations: Array<ShippingMethodTranslation>;
};

export type Query = {
  __typename?: "Query";
  /** The active Channel */
  activeChannel: Channel;
  /** The active Customer */
  activeCustomer?: Maybe<Customer>;
  /**
   * The active Order. Will be `null` until an Order is created via `addItemToOrder`. Once an Order reaches the
   * state of `PaymentAuthorized` or `PaymentSettled`, then that Order is no longer considered "active" and this
   * query will once again return `null`.
   */
  activeOrder?: Maybe<Order>;
  /** Get active payment methods */
  activePaymentMethods: Array<Maybe<PublicPaymentMethod>>;
  /** Get active shipping methods */
  activeShippingMethods: Array<Maybe<PublicShippingMethod>>;
  /** An array of supported Countries */
  availableCountries: Array<Country>;
  /** Returns a Collection either by its id or slug. If neither 'id' nor 'slug' is specified, an error will result. */
  collection?: Maybe<Collection>;
  /** A list of Collections available to the shop */
  collections: CollectionList;
  /** Returns a list of payment methods and their eligibility based on the current active Order */
  eligiblePaymentMethods: Array<PaymentMethodQuote>;
  /** Returns a list of eligible shipping methods based on the current active Order */
  eligibleShippingMethods: Array<ShippingMethodQuote>;
  /** Returns a Facet by its id */
  facet?: Maybe<Facet>;
  /** A list of Facets available to the shop */
  facets: FacetList;
  /** Returns information about the current authenticated User */
  me?: Maybe<CurrentUser>;
  /** Returns the possible next states that the activeOrder can transition to */
  nextOrderStates: Array<Scalars["String"]["output"]>;
  /**
   * Returns an Order based on the id. Note that in the Shop API, only orders belonging to the
   * currently-authenticated User may be queried.
   */
  order?: Maybe<Order>;
  /**
   * Returns an Order based on the order `code`. For guest Orders (i.e. Orders placed by non-authenticated Customers)
   * this query will only return the Order within 2 hours of the Order being placed. This allows an Order confirmation
   * screen to be shown immediately after completion of a guest checkout, yet prevents security risks of allowing
   * general anonymous access to Order data.
   */
  orderByCode?: Maybe<Order>;
  /** Get a Product either by id or slug. If neither 'id' nor 'slug' is specified, an error will result. */
  product?: Maybe<Product>;
  /** Get a list of Products */
  products: ProductList;
  /** Search Products based on the criteria set by the `SearchInput` */
  search: SearchResponse;
};
// ...
export type ProductDetailQuery = {
  __typename?: "Query";
  product?: {
    __typename?: "Product";
    id: string;
    name: string;
    slug: string;
    description: string;
    featuredAsset?: {
      __typename?: "Asset";
      id: string;
      preview: string;
    } | null;
    assets: Array<{ __typename?: "Asset"; id: string; preview: string }>;
    collections: Array<{
      __typename?: "Collection";
      id: string;
      name: string;
      slug: string;
    }>;
    variants: Array<{
      __typename?: "ProductVariant";
      id: string;
      sku: string;
      name: string;
      priceWithTax: number;
      currencyCode: CurrencyCode;
      stockLevel: string;
    }>;
  } | null;
};